Area Of Shape
Area of Shape
Write a program to calculate and display the area of a shape that the user defines via the keyboard. 
1. The program should prompt the user to choose the shape, and then enter the required variables.
For example, if a user chooses a circle shape, then the program should prompt for the radius. Once the user enter the radius, the program calculates and displays the area. 
2. The program should be implemented in any three programming languages of your choice.
3. Do the screenshots of the code and results and upload.